requirements.txt
fastapi
uvicorn[standard]
boto3
python-multipart
Pillow
# Add your digitizer deps here if you use them:
# numpy
# pyembroidery
